TintinEEPROMDownloader: add 5ms delay between chunks.
This is a workaround to ensure DownloaderTest works in Linux.
This is very lightly tested, but has a better chance of working than
before the patch.

The correct fix is to read back each block after write and ensure that it
has been written correctly. This will be implemented in a future update.